Basic damage calc is ready.

+++ include/pokemon.h
@@ -363,7 +363,7 @@ struct BattleMove
     u8 secondaryEffectChance;
     u8 target;
     s8 priority;
-    u8 flags;
+    u32 flags;
     u8 split;
 };
